requirements.txt
Pillow==9.2.0
pytesseract==0.3.10
